- bugfix: prevent remaining files from going below zero
Rationale: Previously, upon each part upload error, remaining files would be reduced by one, leading to messages such as: Completed 4 of 4 parts, -3 file(s) remaining (for a single file). Now, the remaining files are always positive and specify the number of actually failed files. An alternative would be to reduce it to zero and track the failed files in another way, but the method does not seem to make it possible to distinguish between any and last part failed. The ideal fix would be to offer fault tolerance by retrying part uploads.